Fix the user and group resources
The Puppet Dashboard user and group isn't created in the
current Debian packages, which means we need to manage them.
I tried using the same UID and GID as in the Redhat package,
but there were conflicts. As a result, I don't manage the 
UID or GID. Also, the home directory needs to be managed
for the Redhat init script (or else it throws errors).