(11266) Fix values method call on a nil object
This fixes the problem where creating a node without specifying
any classes will try to call the .values method on a
nil object. We now default the klass_hash to an empty hash and return it
regardless of whether classes are passed using the create_nodes method.
This stops the method call on a nil object.